refactor/#65: 피드백 반영중
